% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{approx_metres}
\alias{approx_metres}
\title{Get approximate metres length for a dlat, dlon, works on small scales}
\usage{
approx_metres(dlat, dlon)
}
\description{
Get approximate metres length for a dlat, dlon, works on small scales
}
