% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{build_query_string}
\alias{build_query_string}
\title{Build a db query string for getting geometry within an extent}
\usage{
build_query_string(table_name, extent)
}
\arguments{
\item{extent}{a raster::extent object}
}
\value{
query string
}
\description{
Build a db query string for getting geometry within an extent
}
