% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistance.R
\name{cal_distance_raster}
\alias{cal_distance_raster}
\title{Rasterize some data, and for any NAs, take the distance to the nearest non-NA}
\usage{
cal_distance_raster(data, groundrast)
}
\arguments{
\item{data}{spatial dataframe}

\item{groundrast}{base raster to use}
}
\value{
distance a distance raster
}
\description{
Rasterize some data, and for any NAs, take the distance to the nearest non-NA
}
