% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistance.R
\name{cal_lamp_irradiance}
\alias{cal_lamp_irradiance}
\title{Calculate point irradiance for lamps}
\usage{
cal_lamp_irradiance(lamps, soft_surf, hard_surf, dtm, ext)
}
\arguments{
\item{lamps}{lamp raster}

\item{soft_surf}{soft surface raster}

\item{hard_surf}{hard surface raster}

\item{dtm}{lidar digital terrain map}

\item{ext}{extent (boundary)}
}
\value{
resistance raster
}
\description{
Resistance values in [1, resmax]
}
