% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{cal_resistance_rasters}
\alias{cal_resistance_rasters}
\title{Resistance pipeline: calculate resistance layers which will go into circuitscape}
\usage{
cal_resistance_rasters(
  algorithm_parameters,
  working_dir,
  base_inputs,
  save_images = TRUE
)
}
\arguments{
\item{algorithm_parameters}{an algorithm_parameters object}

\item{working_dir}{directory to save data to}

\item{base_inputs}{input data to the pipeline}
}
\description{
Resistance pipeline: calculate resistance layers which will go into circuitscape
}
