% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistance.R
\name{cal_river_resistance}
\alias{cal_river_resistance}
\title{Calculate a resistance map for rivers}
\usage{
cal_river_resistance(river, groundrast, buffer, resmax, xmax)
}
\arguments{
\item{river}{sp-class (SpatialPoints*, SpatialMultiPoints*, SpatialLines*, or SpatialPolygons*)}

\item{buffer}{meters to river}

\item{resmax}{maximum resistance of "layers"}

\item{xmax}{"slope value"}
}
\value{
raster::raster object for rivers
}
\description{
Rivers are the opposite to roads; they provide conductance.
}
