% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistance.R
\name{cal_road_resistance}
\alias{cal_road_resistance}
\title{Calculate a resistance map for roads}
\usage{
cal_road_resistance(roads, groundrast, buffer, resmax, xmax)
}
\arguments{
\item{roads}{vector data frame}

\item{buffer}{meters to river}

\item{resmax}{maximum resistance of "layers"}

\item{xmax}{"slope value"}
}
\value{
raster::raster object for roads
}
\description{
Roads repel bats and provide resistance; the minimum is zero, when no roads are around.
}
