% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistance.R
\name{calc_surfs}
\alias{calc_surfs}
\title{Given dtm, dsm and building rasters, calculate 'hard' and 'soft' surfaces}
\usage{
calc_surfs(dtm, dsm, buildings)
}
\description{
Hard and soft surfaces are height maps, with 'hard' surfaces being buildings,
and 'soft' surfaces being
}
