% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{call_circuitscape}
\alias{call_circuitscape}
\title{Call circuitscape given a working directory with inputs}
\usage{
call_circuitscape(working_dir, save_images)
}
\arguments{
\item{working_dir}{}

\item{save_images}{bool}
}
\description{
Call circuitscape given a working directory with inputs
}
