% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{connect_to_db}
\alias{connect_to_db}
\title{wrapper function to make unit testing work with mockr}
\usage{
connect_to_db(driver, db_host, db_name, db_port, db_user, db_pass)
}
\description{
wrapper function to make unit testing work with mockr
}
