% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterfunc.R
\name{create_circles}
\alias{create_circles}
\title{Create raster with concentric circles}
\usage{
create_circles(groundrast, x, y, radius, n)
}
\arguments{
\item{groundrast}{base raster to extract circles from}
}
\value{
circles raster
}
\description{
Create raster with concentric circles
}
