% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterfunc.R
\name{create_disk_mask}
\alias{create_disk_mask}
\title{Create a disk at x, y of a given radius, to be used as a mask}
\usage{
create_disk_mask(groundrast, x, y, radius)
}
\arguments{
\item{groundrast}{base raster to extract circles from}
}
\value{
circles raster
}
\description{
Create a disk at x, y of a given radius, to be used as a mask
}
