% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{create_extent}
\alias{create_extent}
\title{Create an extent given algorithm parameters with roost location
 and radius (side length)}
\usage{
create_extent(x, y, delta)
}
\arguments{
\item{delta}{side length of the bounding box}
}
\value{
raster::extent object
}
\description{
Create an extent given algorithm parameters with roost location
 and radius (side length)
}
