% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterfunc.R
\name{create_ground_rast}
\alias{create_ground_rast}
\title{Create a raster for the ground, which is 'NA everywhere except roost coordinates'
 'used for resampling'.}
\usage{
create_ground_rast(x, y, radius, resolution)
}
\arguments{
\item{resolution}{not image resolution, but relative proportion}
}
\value{
raster for ground with roosts
}
\description{
Create a raster for the ground, which is 'NA everywhere except roost coordinates'
 'used for resampling'.
}
