% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{create_raster_query_boundary}
\alias{create_raster_query_boundary}
\title{Convert an extent to a vector of values}
\usage{
create_raster_query_boundary(ext)
}
\arguments{
\item{ext}{a raster::extent object}
}
\value{
boundary a vector of values
}
\description{
Convert an extent to a vector of values
}
