% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistance.R
\name{distance2resistance}
\alias{distance2resistance}
\title{Calculate a resistance map given distance rasters for features}
\usage{
distance2resistance(buffer, rankmax, resmax, xmax, distance_rasters)
}
\arguments{
\item{buffer}{distance from features}

\item{effect}{can be "negative" or something else}
}
\value{
resistance raster
}
\description{
Calculate a resistance map given distance rasters for features
}
