% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing_collection.R
\name{DrawingCollection}
\alias{DrawingCollection}
\title{A DrawingCollection widget allowing the user to draw shapes on a map}
\value{
SpatialPolygonsDataFrame

SpatialLinesDataFrame

SpatialLinesDataFrame

lights dataframe

list of (Spatial) dataframes for each type of drawing
}
\description{
This class, when initialized with a session, input, and map reference,
will insert UI elements, observers, etc. to allow drawing by the user.
This widget is designed to be used within a bsCollapse.

The widget maintains a 'selected' drawing; the one that is selected at any time
is the one that is modified by the user.
}
\examples{
dc <- DrawingCollection$new(input, session, leafletProxy("map"))
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DrawingCollection-new}{\code{DrawingCollection$new()}}
\item \href{#method-DrawingCollection-get_drawings}{\code{DrawingCollection$get_drawings()}}
\item \href{#method-DrawingCollection-select}{\code{DrawingCollection$select()}}
\item \href{#method-DrawingCollection-delete}{\code{DrawingCollection$delete()}}
\item \href{#method-DrawingCollection-write}{\code{DrawingCollection$write()}}
\item \href{#method-DrawingCollection-read_buildings}{\code{DrawingCollection$read_buildings()}}
\item \href{#method-DrawingCollection-read_roads}{\code{DrawingCollection$read_roads()}}
\item \href{#method-DrawingCollection-read_rivers}{\code{DrawingCollection$read_rivers()}}
\item \href{#method-DrawingCollection-read_lights}{\code{DrawingCollection$read_lights()}}
\item \href{#method-DrawingCollection-get_buildings}{\code{DrawingCollection$get_buildings()}}
\item \href{#method-DrawingCollection-get_roads}{\code{DrawingCollection$get_roads()}}
\item \href{#method-DrawingCollection-get_rivers}{\code{DrawingCollection$get_rivers()}}
\item \href{#method-DrawingCollection-get_lights}{\code{DrawingCollection$get_lights()}}
\item \href{#method-DrawingCollection-get_spatial_dfs}{\code{DrawingCollection$get_spatial_dfs()}}
\item \href{#method-DrawingCollection-add_point_complete}{\code{DrawingCollection$add_point_complete()}}
\item \href{#method-DrawingCollection-something_is_selected}{\code{DrawingCollection$something_is_selected()}}
\item \href{#method-DrawingCollection-unselect}{\code{DrawingCollection$unselect()}}
\item \href{#method-DrawingCollection-unselect_no_ui_update}{\code{DrawingCollection$unselect_no_ui_update()}}
\item \href{#method-DrawingCollection-create_new_drawing}{\code{DrawingCollection$create_new_drawing()}}
\item \href{#method-DrawingCollection-clone}{\code{DrawingCollection$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrawingCollection-new"></a>}}
\if{latex}{\out{\hypertarget{method-DrawingCollection-new}{}}}
\subsection{Method \code{new()}}{
read shapefiles


get lines


create a DrawingCollection
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrawingCollection$new(
  input,
  session,
  map,
  ui_selector = "#drawing_collection_ui"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{input}}{shiny input}

\item{\code{session}}{shiny session}

\item{\code{map}}{leaflet proxy}

\item{\code{ui_selector}}{selector for placement}

\item{\code{f}}{folder name}

\item{\code{layer_name}}{name of layer}

\item{\code{type}}{type of drawing to create from the data}

\item{\code{line_mode}}{boolean: if true, create Lines instead of Polygons
Change the type of drawing i to new_type
Add a sequence of points
TODO: unsure where this responsibility should lie
Setup observers for a newly created drawing}

\item{\code{type}}{building type}

\item{\code{use_invisible}}{bool, if FALSE, ignore those that are invisible}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrawingCollection-get_drawings"></a>}}
\if{latex}{\out{\hypertarget{method-DrawingCollection-get_drawings}{}}}
\subsection{Method \code{get_drawings()}}{
get list of drawings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrawingCollection$get_drawings()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrawingCollection-select"></a>}}
\if{latex}{\out{\hypertarget{method-DrawingCollection-select}{}}}
\subsection{Method \code{select()}}{
select drawing i
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrawingCollection$select(i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{integer id for a drawing}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrawingCollection-delete"></a>}}
\if{latex}{\out{\hypertarget{method-DrawingCollection-delete}{}}}
\subsection{Method \code{delete()}}{
delete drawing i
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrawingCollection$delete(i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{integer}}{index i}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrawingCollection-write"></a>}}
\if{latex}{\out{\hypertarget{method-DrawingCollection-write}{}}}
\subsection{Method \code{write()}}{
write shapefiles
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrawingCollection$write(shp_dir)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{shp_dir}}{string directory name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrawingCollection-read_buildings"></a>}}
\if{latex}{\out{\hypertarget{method-DrawingCollection-read_buildings}{}}}
\subsection{Method \code{read_buildings()}}{
read buildings from a file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrawingCollection$read_buildings(f)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f}}{folder with shape files}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrawingCollection-read_roads"></a>}}
\if{latex}{\out{\hypertarget{method-DrawingCollection-read_roads}{}}}
\subsection{Method \code{read_roads()}}{
read roads from a file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrawingCollection$read_roads(f)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f}}{folder with shape files}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrawingCollection-read_rivers"></a>}}
\if{latex}{\out{\hypertarget{method-DrawingCollection-read_rivers}{}}}
\subsection{Method \code{read_rivers()}}{
read rivers from a file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrawingCollection$read_rivers(f)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f}}{folder with shape files}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrawingCollection-read_lights"></a>}}
\if{latex}{\out{\hypertarget{method-DrawingCollection-read_lights}{}}}
\subsection{Method \code{read_lights()}}{
read lights
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrawingCollection$read_lights(f)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{f}}{csv file}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrawingCollection-get_buildings"></a>}}
\if{latex}{\out{\hypertarget{method-DrawingCollection-get_buildings}{}}}
\subsection{Method \code{get_buildings()}}{
get buildings
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrawingCollection$get_buildings(use_invisible = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{use_invisible}}{bool, if FALSE, ignore those that are invisible}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrawingCollection-get_roads"></a>}}
\if{latex}{\out{\hypertarget{method-DrawingCollection-get_roads}{}}}
\subsection{Method \code{get_roads()}}{
get roads
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrawingCollection$get_roads()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrawingCollection-get_rivers"></a>}}
\if{latex}{\out{\hypertarget{method-DrawingCollection-get_rivers}{}}}
\subsection{Method \code{get_rivers()}}{
get rivers
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrawingCollection$get_rivers()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrawingCollection-get_lights"></a>}}
\if{latex}{\out{\hypertarget{method-DrawingCollection-get_lights}{}}}
\subsection{Method \code{get_lights()}}{
get lights
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrawingCollection$get_lights(use_invisible = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrawingCollection-get_spatial_dfs"></a>}}
\if{latex}{\out{\hypertarget{method-DrawingCollection-get_spatial_dfs}{}}}
\subsection{Method \code{get_spatial_dfs()}}{
retrieve spatial data associated with all drawings for use downstream
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrawingCollection$get_spatial_dfs(crs = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{crs}}{an integer CRS (optional)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrawingCollection-add_point_complete"></a>}}
\if{latex}{\out{\hypertarget{method-DrawingCollection-add_point_complete}{}}}
\subsection{Method \code{add_point_complete()}}{
Append a point to a drawing if it is not already marked complete or is not null
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrawingCollection$add_point_complete(x, y, j = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{x coordinate}

\item{\code{y}}{y coordinate}

\item{\code{j}}{optional integer index for the shape to add vector (x, y) to}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrawingCollection-something_is_selected"></a>}}
\if{latex}{\out{\hypertarget{method-DrawingCollection-something_is_selected}{}}}
\subsection{Method \code{something_is_selected()}}{
check if something is selected
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrawingCollection$something_is_selected()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrawingCollection-unselect"></a>}}
\if{latex}{\out{\hypertarget{method-DrawingCollection-unselect}{}}}
\subsection{Method \code{unselect()}}{
unselect currently selected drawing
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrawingCollection$unselect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrawingCollection-unselect_no_ui_update"></a>}}
\if{latex}{\out{\hypertarget{method-DrawingCollection-unselect_no_ui_update}{}}}
\subsection{Method \code{unselect_no_ui_update()}}{
unselect drawing k but do not trigger UI update
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrawingCollection$unselect_no_ui_update(k)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{k}}{integer drawing ID}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrawingCollection-create_new_drawing"></a>}}
\if{latex}{\out{\hypertarget{method-DrawingCollection-create_new_drawing}{}}}
\subsection{Method \code{create_new_drawing()}}{
create a new drawing
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrawingCollection$create_new_drawing(type, height = 10, label = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{string type of drawing}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrawingCollection-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DrawingCollection-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrawingCollection$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
