% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing.R
\docType{class}
\name{DrawnLine}
\alias{DrawnLine}
\title{A line drawn on a map, representing e.g. road}
\value{
sp::Line
}
\description{
A line drawn on a map, representing e.g. road

A line drawn on a map, representing e.g. road
}
\examples{
ls <- DrawnLine$new(1, "road", 10)
}
\section{Super class}{
\code{shiny-bat-dispersion::DrawnShapeBase} -> \code{DrawnLine}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DrawnLine-new}{\code{DrawnLine$new()}}
\item \href{#method-DrawnLine-get_shape}{\code{DrawnLine$get_shape()}}
\item \href{#method-DrawnLine-clone}{\code{DrawnLine$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="shiny-bat-dispersion" data-topic="DrawnShapeBase" data-id="append"><a href='../../shiny-bat-dispersion/html/DrawnShapeBase.html#method-DrawnShapeBase-append'><code>shiny-bat-dispersion::DrawnShapeBase$append()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shiny-bat-dispersion" data-topic="DrawnShapeBase" data-id="append_click_history"><a href='../../shiny-bat-dispersion/html/DrawnShapeBase.html#method-DrawnShapeBase-append_click_history'><code>shiny-bat-dispersion::DrawnShapeBase$append_click_history()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shiny-bat-dispersion" data-topic="DrawnShapeBase" data-id="clear_graphics"><a href='../../shiny-bat-dispersion/html/DrawnShapeBase.html#method-DrawnShapeBase-clear_graphics'><code>shiny-bat-dispersion::DrawnShapeBase$clear_graphics()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shiny-bat-dispersion" data-topic="DrawnShapeBase" data-id="create_ui_element"><a href='../../shiny-bat-dispersion/html/DrawnShapeBase.html#method-DrawnShapeBase-create_ui_element'><code>shiny-bat-dispersion::DrawnShapeBase$create_ui_element()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shiny-bat-dispersion" data-topic="DrawnShapeBase" data-id="get_clicked_xvals"><a href='../../shiny-bat-dispersion/html/DrawnShapeBase.html#method-DrawnShapeBase-get_clicked_xvals'><code>shiny-bat-dispersion::DrawnShapeBase$get_clicked_xvals()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shiny-bat-dispersion" data-topic="DrawnShapeBase" data-id="get_clicked_yvals"><a href='../../shiny-bat-dispersion/html/DrawnShapeBase.html#method-DrawnShapeBase-get_clicked_yvals'><code>shiny-bat-dispersion::DrawnShapeBase$get_clicked_yvals()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shiny-bat-dispersion" data-topic="DrawnShapeBase" data-id="insert_height_param"><a href='../../shiny-bat-dispersion/html/DrawnShapeBase.html#method-DrawnShapeBase-insert_height_param'><code>shiny-bat-dispersion::DrawnShapeBase$insert_height_param()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shiny-bat-dispersion" data-topic="DrawnShapeBase" data-id="set_invisible"><a href='../../shiny-bat-dispersion/html/DrawnShapeBase.html#method-DrawnShapeBase-set_invisible'><code>shiny-bat-dispersion::DrawnShapeBase$set_invisible()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shiny-bat-dispersion" data-topic="DrawnShapeBase" data-id="set_vals"><a href='../../shiny-bat-dispersion/html/DrawnShapeBase.html#method-DrawnShapeBase-set_vals'><code>shiny-bat-dispersion::DrawnShapeBase$set_vals()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shiny-bat-dispersion" data-topic="DrawnShapeBase" data-id="set_visibility"><a href='../../shiny-bat-dispersion/html/DrawnShapeBase.html#method-DrawnShapeBase-set_visibility'><code>shiny-bat-dispersion::DrawnShapeBase$set_visibility()</code></a></span></li>
<li><span class="pkg-link" data-pkg="shiny-bat-dispersion" data-topic="DrawnShapeBase" data-id="set_visible"><a href='../../shiny-bat-dispersion/html/DrawnShapeBase.html#method-DrawnShapeBase-set_visible'><code>shiny-bat-dispersion::DrawnShapeBase$set_visible()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrawnLine-new"></a>}}
\if{latex}{\out{\hypertarget{method-DrawnLine-new}{}}}
\subsection{Method \code{new()}}{
Add the drawing to a map
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrawnLine$new(j, type = "", height = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{j}}{an integer tag}

\item{\code{type}}{an optional type value, e.g. river, or road}

\item{\code{height}}{an optional height value}

\item{\code{map}}{a leaflet map object
Initialize line}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrawnLine-get_shape"></a>}}
\if{latex}{\out{\hypertarget{method-DrawnLine-get_shape}{}}}
\subsection{Method \code{get_shape()}}{
Get a sp Line representation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrawnLine$get_shape()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrawnLine-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DrawnLine-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrawnLine$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
