% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{fetch_vector_inputs}
\alias{fetch_vector_inputs}
\title{Get vector inputs for raster pipeline from db, and combining with inputs}
\usage{
fetch_vector_inputs(algorithm_parameters, working_dir)
}
\arguments{
\item{algorithm_parameters}{an algorithm_parameters object}

\item{working_dir}{directory to save data to}
}
\value{
list of data for input into resistance pipeline
}
\description{
Get vector inputs for raster pipeline from db, and combining with inputs
}
