% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistance.R
\name{filter_binary_layer}
\alias{filter_binary_layer}
\title{Get a function that takes x, y, and returns value if y is 1, otherwise x}
\usage{
filter_binary_layer(value)
}
\arguments{
\item{value}{value to set 1s to}
}
\value{
function f
}
\description{
Get a function that takes x, y, and returns value if y is 1, otherwise x
}
