% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistance.R
\name{get_linear_resistance}
\alias{get_linear_resistance}
\title{Calculate a linear resistance map}
\usage{
get_linear_resistance(drasts, buffer, rankmax, resmax, xmax)
}
\value{
resistance: raster::raster object
}
\description{
Bats like linear features, such as hedgerows; low resistance near the features.
Values in interval [1, resmax]
}
