% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_parameters.R
\name{LampResistance}
\alias{LampResistance}
\title{R6 class representing lamp parameters for resistance map generation}
\description{
Wrapper for buffer, resmax, rankmax, xmax
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LampResistance-new}{\code{LampResistance$new()}}
\item \href{#method-LampResistance-clone}{\code{LampResistance$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LampResistance-new"></a>}}
\if{latex}{\out{\hypertarget{method-LampResistance-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LampResistance$new(resmax, xmax, ext)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LampResistance-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LampResistance-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LampResistance$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
