% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_parameters.R
\name{LandscapeResistance}
\alias{LandscapeResistance}
\title{R6 class representing landscape parameters for resistance map generation}
\description{
Wrapper for resmax, xmax
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LandscapeResistance-new}{\code{LandscapeResistance$new()}}
\item \href{#method-LandscapeResistance-clone}{\code{LandscapeResistance$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LandscapeResistance-new"></a>}}
\if{latex}{\out{\hypertarget{method-LandscapeResistance-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LandscapeResistance$new(rankmax, resmax, xmax)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LandscapeResistance-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LandscapeResistance-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LandscapeResistance$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
