% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistance.R
\name{light_resistance}
\alias{light_resistance}
\title{Calculate light resistance from an irradiance raster}
\usage{
light_resistance(resmax, xmax, irradiance_raster)
}
\description{
Calculate light resistance from an irradiance raster
}
