% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_parameters.R
\name{LinearResistance}
\alias{LinearResistance}
\title{R6 class representing linear parameters for resistance map generation}
\description{
Wrapper for buffer, resmax, rankmax, xmax
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LinearResistance-new}{\code{LinearResistance$new()}}
\item \href{#method-LinearResistance-clone}{\code{LinearResistance$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LinearResistance-new"></a>}}
\if{latex}{\out{\hypertarget{method-LinearResistance-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinearResistance$new(buffer, resmax, rankmax, xmax)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LinearResistance-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LinearResistance-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LinearResistance$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
