% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{log_vector_warnings}
\alias{log_vector_warnings}
\title{Log some warnings if a spatial points object is not as it should be}
\usage{
log_vector_warnings(tag, spdf)
}
\arguments{
\item{tag}{a string tag for warning logging}

\item{spdf}{SpatialPoints or SpatialPointsDataFrame}
}
\description{
Log some warnings if a spatial points object is not as it should be
}
