% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{postprocess_inputs}
\alias{postprocess_inputs}
\title{Get inputs for raster pipeline from db, and combining with inputs}
\usage{
postprocess_inputs(
  algorithm_parameters,
  groundrast,
  vector_inputs,
  raster_inputs,
  working_dir,
  lamps,
  spdfs
)
}
\arguments{
\item{algorithm_parameters}{an algorithm_parameters object}

\item{working_dir}{directory to save data to}

\item{lamps}{a csv file with lamp x, y, z vals}

\item{spdfs}{spatial data frames to combine with db outputs}
}
\value{
list of data for input into resistance pipeline
}
\description{
Get inputs for raster pipeline from db, and combining with inputs
}
