% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistance.R
\name{prep_lidar_rasters}
\alias{prep_lidar_rasters}
\title{Generate new distance rasters}
\usage{
prep_lidar_rasters(surf)
}
\arguments{
\item{surf}{a raster representing the surface}
}
\value{
distance_rasters a matrix of distance rasters
}
\description{
Generate new distance rasters
}
