% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resistance.R
\name{ranked_resistance}
\alias{ranked_resistance}
\title{Calculate resistance from conductance map with ranks}
\usage{
ranked_resistance(conductance, rankmax, resmax, xmax)
}
\description{
Calculate resistance from conductance map with ranks
}
