% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{read_db_raster}
\alias{read_db_raster}
\title{Read a raster from postgis database}
\usage{
read_db_raster(table, ext, db_host, db_name, db_port, db_user, db_pass)
}
\arguments{
\item{ext}{a raster::extent object}
}
\value{
raster::raster
}
\description{
Read a raster from postgis database
}
