% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{read_db_raster_default}
\alias{read_db_raster_default}
\title{Read db raster, return a default if failure, and a boolean flag to indicate failure}
\usage{
read_db_raster_default(
  table,
  ext,
  db_host,
  db_name,
  db_port,
  db_user,
  db_pass,
  default,
  resolution = NULL,
  use_overview = FALSE
)
}
\description{
Read db raster, return a default if failure, and a boolean flag to indicate failure
}
