% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{read_db_vector}
\alias{read_db_vector}
\title{Read vector data from a database}
\usage{
read_db_vector(table_name, ext, db_host, db_name, db_port, db_user, db_pass)
}
\arguments{
\item{extent}{a raster::extent object}
}
\value{
sp-class (SpatialPoints*, SpatialMultiPoints*, SpatialLines*, or SpatialPolygons*)
}
\description{
Read vector data from a database
}
