% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawing_collection.R
\name{remove_height_param}
\alias{remove_height_param}
\title{Remove the UI height param element i}
\usage{
remove_height_param(i)
}
\arguments{
\item{i}{and integer index}
}
\description{
Remove the UI height param element i
}
