% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_parameters.R
\name{RiverResistance}
\alias{RiverResistance}
\title{R6 class representing river parameters for resistance map generation}
\description{
Wrapper for buffer, resmax, xmax
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RiverResistance-new}{\code{RiverResistance$new()}}
\item \href{#method-RiverResistance-clone}{\code{RiverResistance$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiverResistance-new"></a>}}
\if{latex}{\out{\hypertarget{method-RiverResistance-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiverResistance$new(buffer, resmax, xmax)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RiverResistance-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RiverResistance-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RiverResistance$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
