% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_parameters.R
\name{RoadResistance}
\alias{RoadResistance}
\title{R6 class representing road parameters for resistance map generation}
\description{
Wrapper for buffer, resmax, xmax
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RoadResistance-new}{\code{RoadResistance$new()}}
\item \href{#method-RoadResistance-clone}{\code{RoadResistance$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RoadResistance-new"></a>}}
\if{latex}{\out{\hypertarget{method-RoadResistance-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RoadResistance$new(buffer, resmax, xmax)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RoadResistance-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RoadResistance-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RoadResistance$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
