% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm_parameters.R
\name{Roost}
\alias{Roost}
\title{R6 class representing a roost location}
\description{
Wrapper for roost easting, northing, and radius
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Roost-new}{\code{Roost$new()}}
\item \href{#method-Roost-clone}{\code{Roost$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Roost-new"></a>}}
\if{latex}{\out{\hypertarget{method-Roost-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Roost$new(x, y, radius)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Roost-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Roost-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Roost$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
