% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{sp_not_empty}
\alias{sp_not_empty}
\title{Check spatial points object is not empty}
\usage{
sp_not_empty(df)
}
\arguments{
\item{spdf}{SpatialPoints or SpatialPointsDataFrame}
}
\value{
bool
}
\description{
Check spatial points object is not empty
}
